+++ fixtures/index_app/app/controllers/cards/closures_controller.rb
@@ -1,0 +1,9 @@
+class Cards::ClosuresController < ApplicationController
+  def create
+    redirect_to card_path(params[:card_id])
+  end
+
+  def destroy
+    redirect_to card_path(params[:card_id])
+  end
+end