+++ fixtures/index_app/app/jobs/notify_subscribers_job.rb
@@ -1,0 +1,5 @@
+class NotifySubscribersJob < ApplicationJob
+  def perform(article_id)
+    Article.find(article_id).notify_subscribers
+  end
+end