+++ fixtures/index_app/spec/models/article_spec.rb
@@ -1,0 +1,10 @@
+RSpec.describe Article do
+  it "publishes" do
+    article = Article.new
+    article.publish(by: nil)
+
+    expect(article.published?).to be true
+
+    NotifySubscribersJob.perform_now(article.id)
+  end
+end